Tools/cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pbcli"
version = "0.1.0"
description = "Project-Bible shared CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
pb = "pbcli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
